# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: Product defines the structure for an API product
        x-go-package: github.com/F4llenPotAto/golang-microservice-testing/product-api/data
info:
    description: Documentation for Product API
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Return a list of products from the database
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Create a new product
            operationId: createProduct
            parameters:
                - description: |-
                    Product data structure to Update or Create.
                    Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema: {}
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "422":
                    $ref: '#/responses/errorValidation'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        put:
            description: Update a products details
            operationId: updateProduct
            parameters:
                - description: |-
                    Product data structure to Update or Create.
                    Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema: {}
                - description: The id of the product for which the operation relates
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "422":
                    $ref: '#/responses/errorValidation'
            tags:
                - products
    /products/{id}:
        delete:
            description: Update a products details
            operationId: deleteProduct
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        get:
            description: Return a list of products from the database
            operationId: listSingle
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "404":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
produces:
    - application/json
responses:
    errorResponse:
        description: Generic error message returned as a string
    errorValidation:
        description: Validation errors defined as an array of strings
    noContentResponse:
        description: No content is returned by this API endpoint
    productResponse:
        description: Data structure representing a single product
    productsResponse:
        description: A list of products
        schema:
            items: {}
            type: array
schemes:
    - http
swagger: "2.0"